requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
celery>=5.3
redis>=4.6
psycopg2-binary>=2.9
python-dotenv>=1.0
